==> WGCNA/6_wgcna_profile.R <==
# plot expression profile, based on the following repo  

# https://bioinformaticsworkbook.org/dataAnalysis/RNA-Seq/RNA-SeqIntro/wgcna.html#gsc.tab=0

# and plot
# https://bioinformaticsworkbook.org/dataAnalysis/RNA-Seq/RNA-SeqIntro/Assets/wgcna_mod_profile-1.png
